refactor(jazz-next): extract pure inference type operations
