+++ jazz-next/src/JazzNext/Compiler/TypeInference/Capabilities.hs
@@ -18,19 +18,14 @@ module JazzNext.Compiler.TypeInference.Capabilities
     expressionTypeToRuntimeHint,
     finalizeDeferredExplicitConstraintsAt,
     flushCurrentModuleCapabilityFacts,
-    freeTypeVariables,
     freeTypeVariablesInEnv,
-    freeTypeVariablesInTypeSchemeConstraints,
     freshTypeVars,
     importModuleCapabilityFacts,
     inferQualifiedMethodApplication,
     instantiateQualifiedMethodType,
-    instantiateTypeSchemeConstraint,
-    instantiateTypeSchemePrimitiveConstraint,
     mergeCapabilityFacts,
     newInferredClassConstraints,
     qualifiedMethodClassIsVisible,
-    replaceTypeVariables,
     resolveTypeSchemeConstraint,
     restoreCapabilityFacts,
     seedFacts,
@@ -39,7 +34,6 @@ module JazzNext.Compiler.TypeInference.Capabilities
     typeSchemeReferencedCapabilityFacts,
     constraintSignatureTypeToExpressionType,
     constraintSignatureTypeToExpressionTypeWithState,
-    dedupeTypeSchemeConstraints,
     signaturePayloadToSignatureType,
     structuralRuntimeEqualityType,
     updateRootModuleBaselineFacts
@@ -142,6 +136,14 @@ import JazzNext.Compiler.TypeInference.State
     modifyDeclarationState,
     modifyInferenceOutput,
     modifyModuleInferenceState
+  )
+import JazzNext.Compiler.TypeInference.TypeOps
+  ( dedupeTypeSchemeConstraints,
+    freeTypeVariables,
+    freeTypeVariablesInTypeSchemeConstraints,
+    freeTypeVariablesInTypeSchemePrimitiveConstraints,
+    instantiateTypeSchemeConstraint,
+    instantiateTypeSchemePrimitiveConstraint
   )
 import qualified JazzNext.Compiler.TypeInference.Signature as Signature
 import JazzNext.Compiler.TypeInference.Solver
@@ -786,14 +788,6 @@ uniqueExactRuntimeCandidateHint state argumentType candidateHints =
       not (constraintSignatureTypeContainsList candidateHint)
     _ -> False
 
-dedupeTypeSchemeConstraints :: [TypeSchemeConstraint] -> [TypeSchemeConstraint]
-dedupeTypeSchemeConstraints =
-  foldr insertIfMissing []
-  where
-    insertIfMissing constraint constraints
-      | constraint `elem` constraints = constraints
-      | otherwise = constraint : constraints
-
 resolveTypeSchemeConstraint :: InferState -> TypeSchemeConstraint -> TypeSchemeConstraint
 resolveTypeSchemeConstraint state constraint =
   case constraint of
@@ -804,20 +798,6 @@ resolveTypeSchemeConstraint state constraint =
     TypeSchemeMethodConstraint constraintName methodKey argumentType ->
       TypeSchemeMethodConstraint constraintName methodKey (resolveType state argumentType)
 
-freeTypeVariablesInTypeSchemeConstraints :: [TypeSchemeConstraint] -> Set Int
-freeTypeVariablesInTypeSchemeConstraints constraints =
-  Set.unions (map freeTypeVariablesInTypeSchemeConstraint constraints)
-
-freeTypeVariablesInTypeSchemeConstraint :: TypeSchemeConstraint -> Set Int
-freeTypeVariablesInTypeSchemeConstraint constraint =
-  case constraint of
-    TypeSchemeConstraint _ argumentType ->
-      freeTypeVariables argumentType
-    TypeSchemeInferredConstraint _ argumentType ->
-      freeTypeVariables argumentType
-    TypeSchemeMethodConstraint _ _ argumentType ->
-      freeTypeVariables argumentType
-
 resolveTypeSchemePrimitiveConstraint :: InferState -> TypeSchemePrimitiveConstraint -> TypeSchemePrimitiveConstraint
 resolveTypeSchemePrimitiveConstraint state primitiveConstraint =
   case primitiveConstraint of
@@ -825,16 +805,6 @@ resolveTypeSchemePrimitiveConstraint state primitiveConstraint =
       TypeSchemeNumericConstraint numericConstraint (resolveType state argumentType)
     TypeSchemeStrictEqualityConstraint argumentType ->
       TypeSchemeStrictEqualityConstraint (resolveType state argumentType)
-
-freeTypeVariablesInTypeSchemePrimitiveConstraints :: [TypeSchemePrimitiveConstraint] -> Set Int
-freeTypeVariablesInTypeSchemePrimitiveConstraints primitiveConstraints =
-  Set.unions (map freeTypeVariablesInTypeSchemePrimitiveConstraint primitiveConstraints)
-
-freeTypeVariablesInTypeSchemePrimitiveConstraint :: TypeSchemePrimitiveConstraint -> Set Int
-freeTypeVariablesInTypeSchemePrimitiveConstraint primitiveConstraint =
-  case primitiveConstraint of
-    TypeSchemeNumericConstraint _ argumentType -> freeTypeVariables argumentType
-    TypeSchemeStrictEqualityConstraint argumentType -> freeTypeVariables argumentType
 
 freeTypeVariablesInEnv :: InferState -> TypeEnv -> Set Int
 freeTypeVariablesInEnv state =
@@ -875,70 +845,6 @@ freeTypeVariablesInConstructorArgument state argumentType =
       freeTypeVariables (resolveType state expressionType)
     ConstructorArgumentParameter {} -> Set.empty
     ConstructorArgumentFresh -> Set.empty
-
-freeTypeVariables :: ExpressionType -> Set Int
-freeTypeVariables expressionType =
-  case expressionType of
-    TIntType -> Set.empty
-    TIntegerLiteralType {} -> Set.empty
-    TFloatType -> Set.empty
-    TNumericType {} -> Set.empty
-    TBoolType -> Set.empty
-    TCharType -> Set.empty
-    TTextType -> Set.empty
-    TListType elementType ->
-      freeTypeVariables elementType
-    TTupleType elementTypes ->
-      Set.unions (map freeTypeVariables elementTypes)
-    TDataType _ typeArguments ->
-      Set.unions (map freeTypeVariables typeArguments)
-    TFunctionType inputType outputType ->
-      Set.union (freeTypeVariables inputType) (freeTypeVariables outputType)
-    TVarType typeVar ->
-      Set.singleton typeVar
-
-replaceTypeVariables :: Map Int ExpressionType -> ExpressionType -> ExpressionType
-replaceTypeVariables replacements expressionType =
-  case expressionType of
-    TIntType -> TIntType
-    TIntegerLiteralType literalRange -> TIntegerLiteralType literalRange
-    TFloatType -> TFloatType
-    TNumericType numericType -> TNumericType numericType
-    TBoolType -> TBoolType
-    TCharType -> TCharType
-    TTextType -> TTextType
-    TListType elementType ->
-      TListType (replaceTypeVariables replacements elementType)
-    TTupleType elementTypes ->
-      TTupleType (map (replaceTypeVariables replacements) elementTypes)
-    TDataType typeName typeArguments ->
-      TDataType typeName (map (replaceTypeVariables replacements) typeArguments)
-    TFunctionType inputType outputType ->
-      TFunctionType
-        (replaceTypeVariables replacements inputType)
-        (replaceTypeVariables replacements outputType)
-    TVarType typeVar ->
-      Map.findWithDefault expressionType typeVar replacements
-
--- | Pending type signature state mirrors analyzer adjacency rules while
--- carrying the normalized declaration type for the next binding.
-instantiateTypeSchemeConstraint :: Map Int ExpressionType -> TypeSchemeConstraint -> TypeSchemeConstraint
-instantiateTypeSchemeConstraint replacements constraint =
-  case constraint of
-    TypeSchemeConstraint constraintName argumentType ->
-      TypeSchemeConstraint constraintName (replaceTypeVariables replacements argumentType)
-    TypeSchemeInferredConstraint constraintName argumentType ->
-      TypeSchemeInferredConstraint constraintName (replaceTypeVariables replacements argumentType)
-    TypeSchemeMethodConstraint constraintName methodKey argumentType ->
-      TypeSchemeMethodConstraint constraintName methodKey (replaceTypeVariables replacements argumentType)
-
-instantiateTypeSchemePrimitiveConstraint :: Map Int ExpressionType -> TypeSchemePrimitiveConstraint -> TypeSchemePrimitiveConstraint
-instantiateTypeSchemePrimitiveConstraint replacements primitiveConstraint =
-  case primitiveConstraint of
-    TypeSchemeNumericConstraint numericConstraint argumentType ->
-      TypeSchemeNumericConstraint numericConstraint (replaceTypeVariables replacements argumentType)
-    TypeSchemeStrictEqualityConstraint argumentType ->
-      TypeSchemeStrictEqualityConstraint (replaceTypeVariables replacements argumentType)
 
 applyTypeSchemePrimitiveConstraints :: [TypeSchemePrimitiveConstraint] -> InferState -> InferState
 applyTypeSchemePrimitiveConstraints primitiveConstraints state =
